+++ R/PairMarker-package.R
@@ -1,0 +1,9 @@
+#' PairMarker
+#'
+#' Tools for pairwise gene marker discovery using binary phenotype labels.
+#'
+#' @keywords internal
+#' @useDynLib PairMarker, .registration = TRUE
+#' @importFrom Rcpp evalCpp
+#' @importFrom utils read.table write.table
+"_PACKAGE"