+++ R/immunePair-package.R
@@ -1,9 +1,0 @@
-#' immunePair
-#'
-#' Tools for pairwise gene marker analysis in immune response studies.
-#'
-#' @keywords internal
-#' @useDynLib immunePair, .registration = TRUE
-#' @importFrom Rcpp evalCpp
-#' @importFrom utils read.table write.table
-"_PACKAGE"